refactor: herpertext move to shared
